adding maxlen functionality

+++ ft_ls_/libft/incs/ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: tcajee <tcajee@student.wethinkcode.co.za>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/15 14:36:21 by tcajee            #+#    #+#             */
-/*   Updated: 2019/08/30 17:34:13 by tcajee           ###   ########.fr       */
+/*   Updated: 2019/08/30 18:06:00 by tcajee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,10 +81,10 @@ typedef struct group	t_group;
 
 typedef struct			s_form
 {
-	int					grp_len;
-	int					usr_len;
-	int					link_len;
-	int					size_len;
+	size_t				grp_len;
+	size_t				usr_len;
+	size_t				link_len;
+	size_t				size_len;
 }						t_form;
 
 typedef struct			s_info
